pyproject.toml
[project]
name = "simple-jwt-auth"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.8",
    "pydantic>=2.10.6",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]
